Change request parsing to use BufReader
Using BufReader allows us to do read_line(), which means that we can
then split the line on whitespace, getting the method, URI, and
protocol. Once we've extracted those, we can dispatch based on the
results. This makes the code simpler and more robust, while also adding
support for getting the URI of the request.